+++ src/main/java/svenhjol/charmony/core/common/CommonRegistry.java
@@ -6,7 +6,6 @@ import net.minecraft.core.Registry;
 import net.minecraft.core.component.DataComponentType;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.core.registries.BuiltInRegistries;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.world.entity.npc.VillagerProfession;
 import net.minecraft.world.entity.npc.VillagerTrades;
@@ -14,6 +13,8 @@ import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.entity.BlockEntityType;
+import svenhjol.charmony.core.base.Registerable;
+import svenhjol.charmony.core.base.SidedFeature;
 import svenhjol.charmony.core.helper.VillagerHelper;
 
 import java.util.Arrays;
@@ -25,42 +26,44 @@ import static net.minecraft.world.entity.npc.VillagerTrades.WANDERING_TRADER_TRA
 
 @SuppressWarnings("unused")
 public final class CommonRegistry {
-    private static CommonRegistry instance;
+    private final SidedFeature feature;
 
-    public static CommonRegistry instance() {
-        if (instance == null) {
-            instance = new CommonRegistry();
-        }
-        return instance;
+    private CommonRegistry(SidedFeature feature) {
+        this.feature = feature;
     }
 
-    public <B extends Block> Supplier<B> block(ResourceLocation id, B supplier) {
-        return () -> Registry.register(BuiltInRegistries.BLOCK, id, supplier);
+    public static CommonRegistry forFeature(SidedFeature feature) {
+        return new CommonRegistry(feature);
     }
 
-    public <BE extends BlockEntity, B extends Block> Supplier<BlockEntityType<BE>> blockEntity(ResourceLocation id,
+    public <B extends Block> Registerable<B> block(String id, Supplier<B> supplier) {
+        return new Registerable<>(feature, () -> Registry.register(BuiltInRegistries.BLOCK, feature.id(id), supplier.get()));
+    }
+
+    public <BE extends BlockEntity, B extends Block> Registerable<BlockEntityType<BE>> blockEntity(String id,
                                                                                                FabricBlockEntityTypeBuilder.Factory<BE> builder,
                                                                                                List<Supplier<B>> blocks) {
         var blocksToAdd = blocks.stream().map(Supplier::get).toArray(Block[]::new);
 
-        return () -> Registry.register(BuiltInRegistries.BLOCK_ENTITY_TYPE, id,
-            FabricBlockEntityTypeBuilder.create(builder, blocksToAdd).build());
+        return new Registerable<>(feature, () -> Registry.register(BuiltInRegistries.BLOCK_ENTITY_TYPE, feature.id(id),
+            FabricBlockEntityTypeBuilder.create(builder, blocksToAdd).build()));
     }
 
-    public <BE extends BlockEntity> Supplier<BlockEntityType<BE>> blockEntity(ResourceLocation id, FabricBlockEntityTypeBuilder.Factory<BE> builder) {
+    public <BE extends BlockEntity> Registerable<BlockEntityType<BE>> blockEntity(String id, FabricBlockEntityTypeBuilder.Factory<BE> builder) {
         return blockEntity(id, builder, List.of());
     }
 
-    public <D> Supplier<DataComponentType<D>> dataComponent(ResourceLocation id, Supplier<UnaryOperator<DataComponentType.Builder<D>>> dataComponent) {
-        return () -> DataComponents.register(id.toString(), dataComponent.get());
+    public <D> Registerable<DataComponentType<D>> dataComponent(String id, Supplier<UnaryOperator<DataComponentType.Builder<D>>> dataComponent) {
+        return new Registerable<>(feature, () -> DataComponents.register(feature.id(id).toString(), dataComponent.get()));
     }
 
-    public <I extends Item> Supplier<I> item(ResourceLocation id, I supplier) {
-        return () -> Registry.register(BuiltInRegistries.ITEM, id, supplier);
+    public <I extends Item> Registerable<I> item(String id, I supplier) {
+        return new Registerable<>(feature, () -> Registry.register(BuiltInRegistries.ITEM, feature.id(id), supplier));
     }
 
-    public Supplier<SoundEvent> sound(ResourceLocation id) {
-        return () -> Registry.register(BuiltInRegistries.SOUND_EVENT, id, SoundEvent.createVariableRangeEvent(id));
+    public Registerable<SoundEvent> sound(String id) {
+        var res = feature.id(id);
+        return new Registerable<>(feature, () -> Registry.register(BuiltInRegistries.SOUND_EVENT, res, SoundEvent.createVariableRangeEvent(res)));
     }
 
     /**